# Fix performance issues leading to #412