--- a/src/Dhall/TypeCheck.hs
+++ b/src/Dhall/TypeCheck.hs
@@ -63,7 +63,10 @@ axiom :: Const -> Either (TypeError s a) Const
 axiom Type = return Kind
 axiom Kind = Left (TypeError Dhall.Context.empty (Const Kind) Untyped)
 
+
 rule :: Const -> Const -> Either () Const
+-- This forbids dependent types. If this ever changes, then the fast
+-- path in the Let case of typeWithA will become unsound.
 rule Type Kind = Left ()
 rule Type Type = return Type
 rule Kind Kind = return Kind
@@ -155,11 +158,34 @@ typeWithA tpa = loop
                     then return ()
                     else Left (TypeError ctx e (AnnotMismatch a0 nf_A0 nf_A1))
             Nothing -> return ()
+
+        t <- loop ctx _A1
+
         let a1 = Dhall.Core.normalize a0
         let a2 = Dhall.Core.shift 1 (V x 0) a1
-        let b1 = Dhall.Core.subst (V x 0) a2 b0
-        let b2 = Dhall.Core.shift (-1) (V x 0) b1
-        loop ctx b2
+
+        -- The catch-all branch directly implements the Dhall
+        -- specification as written; it is necessary to substitute in
+        -- types in order to get 'dependent let' behaviour and to
+        -- allow type synonyms (see #69). However, doing a full
+        -- substitution is slow if the value is large and used many
+        -- times. If the value being substitued in is a term (i.e.,
+        -- its type is a Type), then we can get a very significant
+        -- speed-up by doing the type-checking once at binding-time,
+        -- as opposed to doing it at every use site (see #412).
+        case Dhall.Core.normalize t of
+          Const Type -> do
+            let ctx' = fmap (Dhall.Core.shift 1 (V x 0)) (Dhall.Context.insert x (Dhall.Core.normalize _A1) ctx)
+            _B0 <- loop ctx' b0
+            let _B1 = Dhall.Core.subst (V x 0) a2 _B0
+            let _B2 = Dhall.Core.shift (-1) (V x 0) _B1
+            return _B2
+
+          _ -> do
+            let b1 = Dhall.Core.subst (V x 0) a2 b0
+            let b2 = Dhall.Core.shift (-1) (V x 0) b1
+            loop ctx b2
+
     loop ctx e@(Annot x t       ) = do
         _ <- loop ctx t