@quasicomputational 2018-06-02
I added the dropped `normalize` and now the branches are obviously quite parallel, and I rewrote the comment and added another to, hopefully, be clearer.