@quasicomputational 2018-06-04
I just pushed another commit that helps matters. Before, on `dhall-to-cabal.dhall`:

```
	   dhall-hash +RTS -p -RTS

	total time  =       15.45 secs   (15446 ticks @ 1000 us, 1 processor)
	total alloc = 8,419,758,564 bytes  (excludes profiling overheads)

COST CENTRE     MODULE                     SRC                                                 %time %alloc

getInput        Text.Megaparsec            Text/Megaparsec.hs:404:1-40                          28.0   21.4
traverseWithKey Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:(420,1)-(421,78)   19.0    8.9
normalizeWith   Dhall.Core                 src/Dhall/Core.hs:(1325,1)-(1675,22)                 13.9   21.5
denote          Dhall.Core                 src/Dhall/Core.hs:(1247,1)-(1307,39)                 12.7   17.3
toHashMap       Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:599:1-44            9.2   10.4
normalize       Dhall.Core                 src/Dhall/Core.hs:1221:1-41                           2.9    6.0
toList          Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:(580,1)-(584,22)    2.6    1.4
shift           Dhall.Core                 src/Dhall/Core.hs:(647,1)-(800,29)                    2.6    3.9
new_            Data.HashMap.Array         Data/HashMap/Array.hs:177:1-28                        2.5    4.7
typeWithA       Dhall.TypeCheck            src/Dhall/TypeCheck.hs:(102,1)-(758,49)               2.2    0.8
```

```
benchmarking issue 412
time                 923.2 ms   (897.1 ms .. 942.8 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 929.3 ms   (924.9 ms .. 932.6 ms)
std dev              4.235 ms   (216.8 μs .. 5.160 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:

```
	   dhall-hash +RTS -p -RTS

	total time  =       12.68 secs   (12684 ticks @ 1000 us, 1 processor)
	total alloc = 7,429,890,836 bytes  (excludes profiling overheads)

COST CENTRE     MODULE                     SRC                                                 %time %alloc

getInput        Text.Megaparsec            Text/Megaparsec.hs:404:1-40                          32.1   24.3
normalizeWith   Dhall.Core                 src/Dhall/Core.hs:(1325,1)-(1675,22)                 15.6   22.9
denote          Dhall.Core                 src/Dhall/Core.hs:(1247,1)-(1307,39)                 15.4   19.6
traverseWithKey Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:(420,1)-(421,78)   15.1    6.8
toHashMap       Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:599:1-44            6.3    7.7
shift           Dhall.Core                 src/Dhall/Core.hs:(647,1)-(800,29)                    3.2    4.4
new_            Data.HashMap.Array         Data/HashMap/Array.hs:177:1-28                        2.8    4.5
normalize       Dhall.Core                 src/Dhall/Core.hs:1221:1-41                           2.0    4.4
toList          Data.HashMap.Strict.InsOrd src/Data/HashMap/Strict/InsOrd.hs:(580,1)-(584,22)    1.8    1.0
typeWithA       Dhall.TypeCheck            src/Dhall/TypeCheck.hs:(102,1)-(757,49)               1.5    0.7
```

```
benchmarking issue 412
time                 893.0 ms   (870.8 ms .. 919.5 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 894.6 ms   (889.8 ms .. 898.3 ms)
std dev              4.839 ms   (3.211 ms .. 5.868 ms)
variance introduced by outliers: 19% (moderately inflated)
```