@quasicomputational 2018-06-04
I'll continue pondering it a bit and see if any more easy optimisations will drop out. `dhall-to-cabal` is back to sort of acceptable performance with these patches (actually better than it's been in a while, despite the 300-constructor union that's been added recently); thanks for the tips and review!