@Gabriel439 2018-06-04
@quasicomputational: You're welcome!  Also, thanks for the performance improvements :)